[fix] - consumer 서비스로직 이동 - 파라미터 오작성으로 수정
